scraping/requirements.txt
attrs==23.2.0
Automat==22.10.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.7
cssselect==1.2.0
defusedxml==0.7.1
dnspython==2.6.1
exceptiongroup==1.2.1
filelock==3.14.0
Flask==3.0.3
geographiclib==2.0
googlemaps==4.10.0
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.2.2
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pymongo==4.7.2
pyOpenSSL==24.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
queuelib==1.7.0
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.2
selenium==4.21.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tldextract==5.1.2
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
typing_extensions==4.11.0
Unidecode==1.3.8
urllib3==2.2.1
w3lib==2.1.2
Werkzeug==3.0.3
wsproto==1.2.0
zope.interface==6.4
